+++ freppledb/input/commands/export.py
@@ -1722,6 +1722,7 @@ class exportItemSuppliers(PlanTask):
                         i.size_multiple,
                         i.size_maximum if i.size_maximum < 10 ** 12 else None,
                         i.batchwindow,
+                        i.extra_safety_leadtime,
                         i.fence,
                         i.cost,
                         i.priority,
@@ -1741,10 +1742,10 @@ class exportItemSuppliers(PlanTask):
                 """
                 insert into itemsupplier
                 (item_id,location_id,supplier_id,effective_start,leadtime,sizeminimum,
-                 sizemultiple,sizemaximum,batchwindow,fence,cost,priority,effective_end,
+                 sizemultiple,sizemaximum,batchwindow,extra_safety_leadtime,fence,cost,priority,effective_end,
                  resource_id,resource_qty,source,lastmodified%s)
                 values(%%s,%%s,%%s,%%s,%%s * interval '1 second',%%s,%%s,%%s,
-                %%s * interval '1 second',%%s * interval '1 second',%%s,%%s,%%s,%%s,%%s,%%s,%%s%s)
+                %%s * interval '1 second',%%s * interval '1 second',%%s * interval '1 second',%%s,%%s,%%s,%%s,%%s,%%s,%%s%s)
                 on conflict (item_id, location_id, supplier_id, effective_start)
                 do update set
                   leadtime=excluded.leadtime,
@@ -1752,6 +1753,7 @@ class exportItemSuppliers(PlanTask):
                   sizemultiple=excluded.sizemultiple,
                   sizemaximum=excluded.sizemaximum,
                   batchwindow=excluded.batchwindow,
+                  extra_safety_leadtime=excluded.extra_safety_leadtime,
                   fence=excluded.fence,
                   cost=excluded.cost,
                   priority=excluded.priority,