+++ src/tags.cpp
@@ -90,6 +90,7 @@ const Keyword Tags::enddate("enddate");
 const Keyword Tags::endtime("endtime");
 const Keyword Tags::entity("entity");
 const Keyword Tags::erase("erase");
+const Keyword Tags::extra_safety_leadtime("extra_safety_leadtime");
 const Keyword Tags::factor("factor");
 const Keyword Tags::feasible("feasible");
 const Keyword Tags::fence("fence");